+++ app/src/main/java/com/davidlcassidy/travelwallet/Database/RefDatabaseHelper.java
@@ -9,10 +9,8 @@ package com.davidlcassidy.travelwallet.Database;
 import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteException;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -21,17 +19,17 @@ import java.io.OutputStream;
 /*
 RefDatabaseHelper is used to manage the Reference Database. It is made up of two
 tables, one for loyalty programs and one for credit cards. These two tables hold the
-static (non user provided) reference data. To maintain data intagrity, this database
+static (non user provided) reference data. To maintain data integrity, this database
 will never have any insert or update queries run within the application and will only
 be updated with new app versions.
  */
 
 public class RefDatabaseHelper extends SQLiteOpenHelper{
 
-	// Database version should be incremented for structural changes to the database.
-	// This will have users re-create their local database on the next app launch.
-	// Unlike the Main Database, updating database version will not result in loss of
-    // user data as long as the unique ID remains unchanged
+	// Database version should be incremented for structural changes to the
+    // database. This will have users re-create their local database on the next app
+    // launch. Unlike the Main Database, updating database version will not result
+    // in loss of user data as long as the unique ID remains unchanged
     public static int DATABASE_VERSION = 1;
 
     private String DB_DIRECTORY;
@@ -41,7 +39,22 @@ public class RefDatabaseHelper extends SQLiteOpenHelper{
     private final Context context;
 
     public static final String TABLE_LP_REF = "loyaltyprograms_ref";
+    public static final String COLUMN_LP_ID = "_id";
+    public static final String COLUMN_LP_TYPEID = "typeId";
+    public static final String COLUMN_LP_NAME = "name";
+    public static final String COLUMN_LP_POINTVALUE = "pointValue";
+    public static final String COLUMN_LP_INACTIVITYEXPIRATION = "inactivityExpiration";
+    public static final String COLUMN_LP_EXPIRATIONOVERRIDE = "expirationOverride";
+    public static final String COLUMN_LP_DEPRECIATED = "depreciated";
+
     public static final String TABLE_CC_REF = "creditcards_ref";
+    public static final String COLUMN_CC_ID = "_id";
+    public static final String COLUMN_CC_BANKID = "bankId";
+    public static final String COLUMN_CC_NAME = "name";
+    public static final String COLUMN_CC_TYPE = "type";
+    public static final String COLUMN_CC_AF = "annualFee";
+    public static final String COLUMN_CC_FTF = "foreignTransactionFee";
+    public static final String COLUMN_CC_DEPRECIATED = "depreciated";
 
     public RefDatabaseHelper(Context context) {
         super(context, DB_NAME, null, 1);