+++ CHANGELOG.md
@@ -1,18 +1,34 @@
-v1.3.0 : 2018-06-24
+v1.3.2 : 2018-06-29
+Feature Release
 
- - Added owners interface
- - Added Chase 5/24 status
+ - Updated database indexing and cursors
+ - Added 13 new credit cards as requested by users
+
+v1.3.1 : 2018-06-25
+Feature Release
+
+ - Updated owners detail activity layout
+ - Updated summary report (minor)
+
+v1.3.0 : 2018-06-24
+Minor Release
+
+ - Added owners interface allowing for multiple program/card owners
+ - Added Chase 5/24 status support
  - Added program/card filters
- - Added program/card notes
+ - Added support for custom program/card notes
  
 v1.2 : 2018-06-15
+Patch Release
 
- - Fixed bug with date fields
+ - Fixed a bug with popup date picker dialog causing stability issues
  
 v1.1 : 2018-06-14
+Patch Release
 
- - Fixed bug with reference database creation
+ - Fixed bug with the reference database creation causing stability issues
  
 v1.0 : 2018-06-09
+Major Release
 
  - Initial release